[CHORE] ViewController에 남아있는 GuideView 코드 정리(#399)
